Add git hook to process changelog
